Abstract backend p1
